+++ core/src/ee/taltech/deepdarkdungeon/Models/BadCgaracterClasses/SkeletonWarrior.java
@@ -5,7 +5,7 @@ import ee.taltech.deepdarkdungeon.Models.GameObject;
 
 public class SkeletonWarrior extends GameObject {
 
-    public SkeletonWarrior (Texture picture, String name, int health, int power, double x, double y, double width, double height, BadCharacterClass badCharacterClass, CharacterType badCharacterType) {
+    public SkeletonWarrior (Texture picture, String name, int health, int power, int x, int y, double width, double height, BadCharacterClass badCharacterClass, CharacterType badCharacterType) {
         this.name = name;
         this.health = health;
         this.power = power;
