+++ core/src/ee/taltech/deepdarkdungeon/Models/BadCgaracterClasses/Zombie.java
@@ -6,7 +6,7 @@ import ee.taltech.deepdarkdungeon.Models.GameObject;
 public class Zombie extends GameObject {
 
 
-    public Zombie (Texture picture, String name, int health, int power, double x, double y, double width, double height, BadCharacterClass badCharacterClass, CharacterType badCharacterType) {
+    public Zombie (Texture picture, String name, int health, int power, int x, int y, double width, double height, BadCharacterClass badCharacterClass, CharacterType badCharacterType) {
         this.name = name;
         this.health = health;
         this.power = power;
