+++ core/src/ee/taltech/deepdarkdungeon/Models/characterClasses/Archer.java
@@ -5,7 +5,7 @@ import ee.taltech.deepdarkdungeon.Models.GameObject;
 
 public class Archer extends GameObject {
 
-    public Archer(Texture picture, String name, int health, int power, double x, double y, double width, double height, CharacterClass characterClass, CharacterType characterType) {
+    public Archer(Texture picture, String name, int health, int power, int x, int y, double width, double height, CharacterClass characterClass, CharacterType characterType) {
         this.name = name;
         this.health = health;
         this.mana = 100;
