+++ core/src/ee/taltech/deepdarkdungeon/Models/characterClasses/Magic.java
@@ -5,7 +5,7 @@ import ee.taltech.deepdarkdungeon.Models.GameObject;
 
 public class Magic extends GameObject {
 
-    public Magic(Texture picture, String name, int health, int power, double x, double y, double width, double height, CharacterClass characterClass, CharacterType characterType) {
+    public Magic(Texture picture, String name, int health, int power, int x, int y, double width, double height, CharacterClass characterClass, CharacterType characterType) {
         this.name = name;
         this.health = health;
         this.power = power;
